requirements.txt
pytest==8.3.4
requests==2.32.3
playwright==1.50.0
pytest-playwright==0.7.0
pytest-asyncio==0.25.3
python-dotenv==1.0.1
